@@ -8,13 +8,12 @@ using Microsoft.ApplicationInsights.AspNetCore.Extensions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Diagnostics.HealthChecks;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
-using Newtonsoft.Json.Serialization;
+using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
 using Polly;
-using Swashbuckle.AspNetCore.Swagger;
 
 namespace Azure_AD_Users_Extract
 {
@@ -47,20 +46,25 @@ namespace Azure_AD_Users_Extract
                 .AddTransientHttpErrorPolicy(builder =>
                     builder.WaitAndRetryAsync(2, _ => TimeSpan.FromMilliseconds(500)));
 
-            services.AddMvc(options => { options.Filters.Add<ExceptionActionFilter>(); })
-                .AddJsonOptions(options =>
-                {
-                    var resolver  = options.SerializerSettings.ContractResolver;
-                    var res = (DefaultContractResolver) resolver;
-                    if (res != null)
-                    {
-                        res.NamingStrategy = null; // <-- this removes the default camelCasing of object property names when serializing to Json
-                    }
-                })
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            //services.AddMvc(options => { options.Filters.Add<ExceptionActionFilter>(); })
+            //    .AddJsonOptions(options =>
+            //    {
+            //        var resolver  = options.SerializerSettings.ContractResolver;
+            //        var res = (DefaultContractResolver) resolver;
+            //        if (res != null)
+            //        {
+            //            res.NamingStrategy = null; // <-- this removes the default camelCasing of object property names when serializing to Json
+            //        }
+            //    })
+            //    .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+
+            services.AddControllers(configure =>
+            {
+                configure.Filters.Add<ExceptionActionFilter>(); 
+            });
 
             services.AddSwaggerGen(c => {  
-                c.SwaggerDoc("v1", new Info {  
+                c.SwaggerDoc("v1", new OpenApiInfo {  
                     Version = "v1",  
                     Title = "Azure_AD_Users_Extract API",  
                     Description = "Azure_AD_Users_Extract ASP.NET Core Web API"  
@@ -77,7 +81,7 @@ namespace Azure_AD_Users_Extract
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -89,15 +93,20 @@ namespace Azure_AD_Users_Extract
                 app.UseHsts();
             }
 
-            app.UseHealthChecks("/healthcheck", new HealthCheckOptions
-            {
-                Predicate = _ => true,
-                ResponseWriter = UIResponseWriter.WriteHealthCheckUIResponse
-            });
-
             app.UseHttpsRedirection();
             app.UseStatusCodePages();
-            app.UseMvc();
+            app.UseRouting();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapHealthChecks("/health", new HealthCheckOptions
+                {
+                    Predicate = _ => true,
+                    ResponseWriter = UIResponseWriter.WriteHealthCheckUIResponse
+                });
+
+                endpoints.MapControllers();
+            });
+
             app.UseSwagger();
             app.UseSwaggerUI(c =>
             {
