@@ -3,7 +3,7 @@ using Azure_AD_Users_Publisher.Services.Models;
 
 namespace Azure_AD_Users_Publisher.Services
 {
-    public interface ISalesforceUserPublishService
+    public interface ISalesforceUserService
     {
         int ErrorCount { get; }
 
